+++ spritesheet_functions.py
@@ -15,3 +15,20 @@ class SpriteSheet(object):
         image.set_colorkey((0,0,0))
 
         return image
+
+
+class Wall(object):
+    """it walls
+
+    attribute: rect"""
+    def __init__(self, pos):
+        walls.append(self)
+        self.rect = pygame.Rect(pos[0], pos[1], 32, 32)
+
+class ExitBlock(object):
+    """It exits
+
+    attribute: rect"""
+    def __init__(self, pos):
+        exit_blocks.append(self)
+        self.rect = pygame.Rect(pos[0], pos[1], 32, 32)
